+++ plugins/google/helpers/requester.ts
@@ -112,12 +112,10 @@ const googleRequest: IGoogleRequest = {
         .get<U>(url, {
           headers: getHeaders,
         })
-        .then((response) => {
-          const resp = handleResponse<U>(response);
-          resolve(resp);
-        })
+        .then((response) => handleResponse<U>(response))
+        .then(resolve)
         .catch((error: AxiosError) => {
-          reject(error);
+          reject(new Error(`${error.name}: ${error.message}`));
         });
     }),
 
@@ -149,12 +147,10 @@ const googleRequest: IGoogleRequest = {
         .post<V>(url, data, {
           headers: getHeaders,
         })
-        .then((response) => {
-          const resp = handleResponse<V>(response);
-          resolve(resp);
-        })
+        .then((response) => handleResponse<V>(response))
+        .then(resolve)
         .catch((error: AxiosError) => {
-          reject(error);
+          reject(new Error(`${error.name}: ${error.message}`));
         });
     }),
 
@@ -186,12 +182,10 @@ const googleRequest: IGoogleRequest = {
         .patch<V>(url, data, {
           headers: getHeaders,
         })
-        .then((response) => {
-          const resp = handleResponse<V>(response);
-          resolve(resp);
-        })
+        .then((response) => handleResponse<V>(response))
+        .then(resolve)
         .catch((error: AxiosError) => {
-          reject(error);
+          reject(new Error(`${error.name}: ${error.message}`));
         });
     }),
 
@@ -218,12 +212,10 @@ const googleRequest: IGoogleRequest = {
           headers: getHeaders,
           data,
         })
-        .then((response) => {
-          const resp = handleResponse<U>(response);
-          resolve(resp);
-        })
+        .then((response) => handleResponse<U>(response))
+        .then(resolve)
         .catch((error: AxiosError) => {
-          reject(error);
+          reject(new Error(`${error.name}: ${error.message}`));
         });
     }),