+++ plugins/google/helpers/resolve-token.ts
@@ -9,7 +9,7 @@ import { objectID } from '@plugins/misc';
 // Types
 import type { TGoogleApiScope } from '@plugins/google/helpers/types';
 import type { ICredentials, ICredentialsDoc } from '@models/credential/types';
-import type { IToken, ITokenDoc } from '@models/tokens/types';
+import type { ITokenDoc } from '@models/tokens/types';
 import type { IServiceAccDoc } from '@models/service-account/types';
 import type { Error as MongoError } from 'mongoose';
 import type {
@@ -17,78 +17,82 @@ import type {
   IGetAllTokens,
 } from '@plugins/google/helpers/types';
 
+const handleCredential = (
+  credential: ICredentialsDoc | null,
+  scopes: TGoogleApiScope[],
+): Promise<IGetAllTokens> =>
+  new Promise<IGetAllTokens>((resolve, reject) => {
+    const response: IGetAllTokens = {
+      success: false,
+    };
+    if (credential) {
+      response.credential = credential;
+      ServiceAccs.find({ related_to: credential._id })
+        .then((serviceAccs) => {
+          response.serviceAcc = serviceAccs;
+          const serviceAccountIds = serviceAccs.map((account) => account._id);
+          const tokenFindParam = {
+            $or: [
+              { related_to: credential._id, scopes },
+              ...serviceAccountIds.map((id) => ({ related_to: id, scopes })),
+            ],
+          };
+          return tokenFindParam;
+        })
+        .then((searchParam) => Tokens.find(searchParam))
+        .then((tokens) => {
+          if (tokens.length > 0) {
+            const normalAccessTokens = tokens.filter(
+              (token) =>
+                token.type === 'access' && token.ref_model === 'Credential',
+            );
+            const serviceAccessTokens = tokens.filter(
+              (token) =>
+                token.type === 'access' && token.ref_model === 'ServiceAccount',
+            );
+            const refreshTokens = tokens.filter(
+              (token) => token.type === 'refresh',
+            );
+            response.tokens = {
+              access: {
+                normal: normalAccessTokens,
+                service: serviceAccessTokens,
+              },
+              refresh: refreshTokens,
+            };
+            response.success = true;
+            return response;
+          } else {
+            response.success = false;
+            return response;
+          }
+        })
+        .then(resolve)
+        .catch((error: MongoError) => {
+          reject(new Error(`${error.name}: ${error.message}`));
+        });
+    } else {
+      reject(new Error('Credentials ID Not Found in the Database'));
+    }
+  });
+
 /**
  * Fetches all the Data in the Database Related to a Credential ID
  *
- * @param { string } credentialID - Credential ID from the Database
+ * @param {string} credentialID - Credential ID from the Database
  * @param {TGoogleApiScope[]} scopes - Google Oauth API Scopes
- * @returns { IGetAllTokens } - Data Related to Credential ID
+ * @returns {IGetAllTokens} - Data Related to Credential ID
  */
 function getAllTokens(
   credentialID: ICredentials['_id'],
   scopes: TGoogleApiScope[],
 ): Promise<IGetAllTokens> {
   return new Promise<IGetAllTokens>((resolve, reject) => {
-    const response: IGetAllTokens = {
-      success: false,
-    };
     Credentials.findById(credentialID)
-      .then((credential) => {
-        if (credential) {
-          response.credential = credential;
-          ServiceAccs.find({ related_to: credential._id })
-            .then((serviceAccs) => {
-              response.serviceAcc = serviceAccs;
-              const serviceAccountIds = serviceAccs.map(
-                (account) => account._id,
-              );
-              const tokenFindParam = [
-                { related_to: credential._id, scopes },
-                ...serviceAccountIds.map((id) => ({ related_to: id, scopes })),
-              ];
-              Tokens.find({
-                $or: tokenFindParam,
-              })
-                .then((tokens) => {
-                  if (tokens.length > 0) {
-                    const normalAccessTokens = tokens.filter(
-                      (token) =>
-                        token.type === 'access' &&
-                        token.ref_model === 'Credential',
-                    );
-                    const serviceAccessTokens = tokens.filter(
-                      (token) =>
-                        token.type === 'access' &&
-                        token.ref_model === 'ServiceAccount',
-                    );
-                    const refreshTokens = tokens.filter(
-                      (token) => token.type === 'refresh',
-                    );
-                    response.tokens = {
-                      access: {
-                        normal: normalAccessTokens,
-                        service: serviceAccessTokens,
-                      },
-                      refresh: refreshTokens,
-                    };
-                    response.success = true;
-                    resolve(response);
-                  } else {
-                    response.success = false;
-                    resolve(response);
-                  }
-                })
-                .catch((error: MongoError) => {
-                  reject(new Error(`${error.name}: ${error.message}`));
-                });
-            })
-            .catch((error: MongoError) => {
-              reject(new Error(`${error.name}: ${error.message}`));
-            });
-        }
-      })
-      .catch((error: MongoError) => {
-        reject(new Error(`${error.name}: ${error.message}`));
+      .then((credential) => handleCredential(credential, scopes))
+      .then(resolve)
+      .catch((err: string) => {
+        reject(new Error(err));
       });
   });
 }
@@ -152,28 +156,20 @@ function generateNormalTokenSave(
 ): Promise<ITokenDoc> {
   return new Promise<ITokenDoc>((resolve, reject) => {
     generateNormalAccessToken(credentials, refreshToken.token)
-      .then((response) => {
-        const uid = objectID('t');
-        const now = Date.now();
-        const token: IToken = {
-          _id: uid,
-          token: response.access_token,
-          type: 'access',
-          related_to: credentials._id,
-          scopes,
-          ref_model: 'Credential',
-          expires_at: now + response.expires_in * 1000,
-          website: 'google.com',
-        };
-        Tokens.create(token)
-          .then((savedToken) => {
-            resolve(savedToken);
-          })
-          .catch((err: MongoError) => {
-            reject(new Error(`${err.name}: ${err.message}`));
-          });
-      })
-      .catch((err) => {
+      .then((response) => Promise.all([response, objectID('t'), Date.now()]))
+      .then(([response, uid, now]) => ({
+        _id: uid,
+        token: response.access_token,
+        type: 'access',
+        related_to: credentials._id,
+        scopes,
+        ref_model: 'Credential',
+        expires_at: now + response.expires_in * 1000,
+        website: 'google.com',
+      }))
+      .then((token) => Tokens.create(token))
+      .then(resolve)
+      .catch((err: string) => {
         reject(new Error(err));
       });
   });
@@ -192,28 +188,20 @@ function generateServiceTokenSave(
 ): Promise<ITokenDoc> {
   return new Promise<ITokenDoc>((resolve, reject) => {
     generateServiceAccessToken(account, scopes)
-      .then((response) => {
-        const uid = objectID('t');
-        const now = Date.now();
-        const token: IToken = {
-          _id: uid,
-          token: response.access_token,
-          type: 'access',
-          related_to: account._id,
-          scopes,
-          ref_model: 'ServiceAccount',
-          expires_at: now + response.expires_in * 1000,
-          website: 'google.com',
-        };
-        Tokens.create(token)
-          .then((savedToken) => {
-            resolve(savedToken);
-          })
-          .catch((err: MongoError) => {
-            reject(new Error(`${err.name}: ${err.message}`));
-          });
-      })
-      .catch((err) => {
+      .then((response) => Promise.all([response, objectID('t'), Date.now()]))
+      .then(([response, uid, now]) => ({
+        _id: uid,
+        token: response.access_token,
+        type: 'access',
+        related_to: account._id,
+        scopes,
+        ref_model: 'ServiceAccount',
+        expires_at: now + response.expires_in * 1000,
+        website: 'google.com',
+      }))
+      .then((token) => Tokens.create(token))
+      .then(resolve)
+      .catch((err: string) => {
         reject(new Error(err));
       });
   });
@@ -293,9 +281,9 @@ function serviceAccountTokenHandler(
 /**
  * Checks Validity of Access Tokens and Refreshes it
  *
- * @param { IGetAllTokens } tokenData - Response from GetallTokens Function
- * @param { TGoogleApiScope[] } scopes - Google Oauth API Scopes
- * @returns { IGetAllTokens } - Active Tokens
+ * @param {IGetAllTokens} tokenData - Response from GetallTokens Function
+ * @param {TGoogleApiScope[]} scopes - Google Oauth API Scopes
+ * @returns {IGetAllTokens} - Active Tokens
  */
 function checkTokenRefreshit(
   tokenData: IGetAllTokens,
@@ -383,9 +371,9 @@ function checkTokenRefreshit(
 /**
  * Resolves a Access Token for the Respective Google Credential ID (Incl. Service Accounts)
  *
- * @param { string } credentialID - Credentials ID From Database
- * @param { TGoogleApiScope[] } scopes - Google OAuth API Scopes
- * @returns { Promise<ITokenResolver> } - Resolves a Token to use in Google API
+ * @param {string} credentialID - Credentials ID From Database
+ * @param {TGoogleApiScope[]} scopes - Google OAuth API Scopes
+ * @returns {Promise<ITokenResolver>} - Resolves a Token to use in Google API
  */
 export default function (
   credentialID: ICredentials['_id'],